+++ libs/common/src/lib/interfaces/portfolio-summary.interface.ts
@@ -1,0 +1,12 @@
+import { PortfolioPerformance } from './portfolio-performance.interface';
+
+export interface PortfolioSummary extends PortfolioPerformance {
+  cash: number;
+  committedFunds: number;
+  fees: number;
+  firstOrderDate: Date;
+  netWorth: number;
+  ordersCount: number;
+  totalBuy: number;
+  totalSell: number;
+}